% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oe.anomaly.r
\name{oe_anomaly}
\alias{oe_anomaly}
\title{Observed to Expected (O/E) Anomaly Detection}
\usage{
oe_anomaly(df, obs_column, exp_column, target_column)
}
\arguments{
\item{df}{input data frame with column of observed values and expected values}

\item{obs_column}{Argument that specifies the column in data frame with observed values}

\item{exp_column}{Argument that specifies the column in data frame with expected (predicted) values}

\item{target_column}{Argument that specifies the column in data frame with targets (leave argument blank if no targets available)}
}
\value{
This function returns a data frame with O/E ratio control limits columns and violation indicators (anomalies) as well as target method ("center" -> O/E Ratio of 1 OR "Target" -> selected target) **NOTE: Expected Values must be greater than ZERO.
}
\description{
Statistical significance and confidence intervals for an Observed to Expected Ratio
Vandenbroucke (1982) method.

Formula Reference: https://rpubs.com/lefkios_paikousis/smr
}
\examples{
\dontrun{

GROUP <- c("A", "B", "C", "D", "E", "F")
NUMERATOR_NO <- c(2, 6, 10, 2, 5, 0)
DENOMINATOR_NO <- c(3, 3.3, 2, 0.5, 8, 10)
TARGET_NO <- c(0.8, 0.5, 0.8, 0.8, 0.7, NA)

df_test <- data.frame(GROUP, NUMERATOR_NO, DENOMINATOR_NO, TARGET_NO )

 df_out <- oe_anomaly(df_test, "NUMERATOR_NO", "DENOMINATOR_NO", "TARGET_NO" )
 


 }
}
